add compatibility mode metric
