+++ internal/reconciler/telemetry/reconciler.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/prometheus/client_golang/prometheus"
 	"gopkg.in/yaml.v3"
 	admissionregistrationv1 "k8s.io/api/admissionregistration/v1"
 	corev1 "k8s.io/api/core/v1"
@@ -30,11 +29,6 @@ const (
 	selfMonitorConfigFileName    = "prometheus.yml"
 	selfMonitorAlertRuleFileName = "alerting_rules.yml"
 )
-
-var compatibilityModeGauge = prometheus.NewGauge(prometheus.GaugeOpts{
-	Name: "telemetry_module_otelcol_metrics_compatibility_mode",
-	Help: "Indicates if the OpenTelemetry internal metrics compatibility mode is enabled (1) or disabled (0)",
-})
 
 type Config struct {
 	Traces      TracesConfig
@@ -135,7 +129,7 @@ func (r *Reconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Resu
 		}
 	}
 
-	updateCompatibilityModeMetric(ctx, r.Client, telemetry.Namespace)
+	updateCompatibilityModeMetric(telemetryutils.GetCompatibilityModeFromTelemetry(ctx, r.Client, telemetry.Namespace))
 
 	requeue := telemetry.Status.State == operatorv1alpha1.StateWarning
 
@@ -303,13 +297,4 @@ func (r *Reconciler) reconcileWebhook(ctx context.Context, telemetry *operatorv1
 	}
 
 	return nil
-}
-
-func updateCompatibilityModeMetric(ctx context.Context, c client.Client, namespace string) {
-	compatibilityMode := telemetryutils.GetCompatibilityModeFromTelemetry(ctx, c, namespace)
-	if compatibilityMode {
-		compatibilityModeGauge.Set(1)
-	} else {
-		compatibilityModeGauge.Set(0)
-	}
 }
